predict: move market bettability guard to utils
